requirements.txt
certifi==2020.12.5
chardet==4.0.0
docx2txt==0.8
idna==2.10
numpy==1.20.1
openai==0.4.0
python-dotenv==0.15.0
requests==2.25.1
urllib3==1.26.3
